msm: rq_stats: Calculate load based on current freq limit
We are not using rq_stats any more so this commit is just for the
sake of completeness. Prior to this patch, load has been calculated
based on an initial value of policy.cpuinfo.max_freq. This value may
change for several reasons: userspace settings, thermal throttling
or extended frequency tables where the absolute maximum level
is often not used during regular operations. This results
in unreliable calculations. Consider current policy.max
to fix this behaviour.

Signed-off-by: Zdrowy Gosciu <ZdrowyGosciu+GITHUB@gmail.com>